+++ packages/core/src/utils/styles/index.ts
@@ -81,6 +81,7 @@ export * from '../pagination-utils'
 export * from '../dropdown-utils'
 export * from '../context-menu-utils'
 export * from '../navigation-menu-utils'
+export * from '../page-header-utils'
 
 // Feedback components
 export * from '../drawer-utils'